requirements.txt
numpy
biopython
#conda
pandas
seaborn
chardet
argparse
multiprocess
#subprocess
#shutil
#glob
joblib
ujson
pyyaml
matplotlib
#pkgutil